TIP-303: move JobFactory and StepFactory from the batch bundle to the component
