+++ src/Akeneo/Bundle/BatchBundle/Connector/ConnectorRegistry.php
@@ -2,8 +2,8 @@
 
 namespace Akeneo\Bundle\BatchBundle\Connector;
 
-use Akeneo\Bundle\BatchBundle\Job\JobFactory;
-use Akeneo\Bundle\BatchBundle\Step\StepFactory;
+use Akeneo\Component\Batch\Job\JobFactory;
+use Akeneo\Component\Batch\Step\StepFactory;
 use Akeneo\Component\Batch\Job\Job;
 use Akeneo\Component\Batch\Job\JobInterface;
 use Akeneo\Component\Batch\Model\JobInstance;
